# creditsoft/openapi.yaml
openapi: 3.1.0
info:
  title: CreditSoft Local Partner API
  version: 1.0.0
  description: |
    Local-first API for trusted lead-management systems and the CreditSoft website portal.
    Authenticate every write and read request with a bearer token generated inside the intranet settings.
servers:
  - url: /api/v1
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: CreditSoftToken
  schemas:
...
    ClientCreate:
      type: object
      required:
        - first_name
        - last_name
      properties:
        first_name:
          type: string
        last_name:
          type: string
        email:
          type: string
          format: email
        phone:
          type: string
        address_line_1:
          type: string
        address_line_2:
          type: string
        city:
          type: string
        state:
          type: string
        postal_code:
          type: string
        current_score:
          type: integer
          minimum: 300
          maximum: 850
        status:
          type: string
          example: lead
        goals:
          type: string
        external_reference:
          type: string
        metadata:
          type: object
          additionalProperties: true
...
    ClientNoteCreate:
      type: object
      required:
        - visibility
        - note
      properties:
        reporting_cycle_id:
          type:
            - integer
            - "null"
        visibility:
          type: string
          enum:
            - private_note
            - working_note
            - shareable_case_brief
        note:
          type: string
        ai_summary:
          type:
            - string
            - "null"
    ClientTaskCreate:
      type: object
      required:
        - title
        - priority
      properties:
        reporting_cycle_id:
          type:
            - integer
            - "null"
        title:
          type: string
        details:
          type:
            - string
            - "null"
        priority:
          type: string
          enum:
            - low
            - normal
            - high
        due_at:
          type:
            - string
            - "null"
          format: date-time
...
    BrowserCaptureCreate:
      type: object
      required:
        - reporting_cycle_id
      properties:
        reporting_cycle_id:
          type: integer
        source_type:
          type:
            - string
            - "null"
          enum:
            - dom_capture
            - browser_capture
            - extension_capture
            - safari_webarchive
        browser_name:
          type:
            - string
            - "null"
        page_title:
          type:
            - string
            - "null"
        page_url:
          type:
            - string
            - "null"
        html:
          type:
            - string
            - "null"
...
paths:
...
  /clients:
    post:
      summary: Create a local lead/client record
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ClientCreate'
      responses:
        "201":
          description: Client created
...
  /clients/{clientCuid}:
    get:
      summary: Read the local client profile for a client
      parameters:
        - in: path
          name: clientCuid
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Client profile
    patch:
      summary: Update local client contact fields and case basics with audit logging
      parameters:
        - in: path
          name: clientCuid
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ClientCreate'
      responses:
        "200":
          description: Client updated
...
  /clients/{clientCuid}/cycles:
    get:
      summary: List reporting cycles for a client
      parameters:
        - in: path
          name: clientCuid
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Reporting cycles
...
  /clients/{clientCuid}/notes:
    get:
      summary: List client notes
      parameters:
        - in: path
          name: clientCuid
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Notes list
    post:
      summary: Add a client note with audit logging
      parameters:
        - in: path
          name: clientCuid
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ClientNoteCreate'
      responses:
        "201":
          description: Note created
  /clients/{clientCuid}/violations:
    get:
      summary: List violation candidates for a client
      parameters:
        - in: path
          name: clientCuid
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Violations list
  /clients/{clientCuid}/letters:
    get:
      summary: List dispute letters for a client
      parameters:
        - in: path
          name: clientCuid
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Letters list
  /clients/{clientCuid}/briefs:
    get:
      summary: List shareable case briefs for a client
      parameters:
        - in: path
          name: clientCuid
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Briefs list
  /clients/{clientCuid}/tasks:
    get:
      summary: List client tasks
      parameters:
        - in: path
          name: clientCuid
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Tasks list
    post:
      summary: Create a client task with audit logging
      parameters:
        - in: path
          name: clientCuid
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ClientTaskCreate'
      responses:
        "201":
          description: Task created
...
  /clients/{clientCuid}/browser-captures:
    get:
      summary: List browser captures uploaded for a client
      parameters:
        - in: path
          name: clientCuid
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Browser captures list
    post:
      summary: Upload browser evidence or a Safari webarchive with audit logging
      parameters:
        - in: path
          name: clientCuid
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/BrowserCaptureCreate'
          application/json:
            schema:
              $ref: '#/components/schemas/BrowserCaptureCreate'
      responses:
        "201":
          description: Browser capture created
...